requirements.txt
a2wsgi==1.10.10
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosmtplib==5.1.0
aiosqlite==0.21.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
apache-airflow==2.7.2
apache-airflow-providers-common-sql==1.15.0
apache-airflow-providers-ftp==3.10.1
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-sqlite==3.8.2
apispec==6.10.0
argcomplete==3.6.3
asgiref==3.11.1
attrs==26.1.0
babel==2.18.0
blinker==1.9.0
cachelib==0.13.0
cadwyn==5.4.6
cattrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==3.3.0
cron_descriptor==2.0.8
croniter==6.2.2
cryptography==46.0.6
Deprecated==1.3.1
dill==0.4.1
dnspython==2.8.0
email-validator==1.3.1
fastapi==0.117.1
fastapi-cli==0.0.24
fiona==1.10.1
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozenlist==1.8.0
fsspec==2026.3.0
geopandas==1.1.3
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-crc32c==1.8.0
google-re2==1.1.20251105
google-resumable-media==2.8.0
googleapis-common-protos==1.73.1
graphviz==0.21
greenback==1.3.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.78.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kafka-python==2.3.0
kafka-python-ng==2.2.3
lazy-object-proxy==1.12.0
libcst==1.8.6
limits==5.8.0
linkify-it-py==2.1.0
lockfile==0.12.2
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
marshmallow-oneofschema==3.2.0
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.8.0
msgspec==0.20.0
multidict==6.7.1
natsort==8.4.0
networkx==3.6.1
numpy==2.4.3
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
ordered-set==4.1.0
osmnx==2.1.0
outcome==1.3.0.post0
packaging==26.0
pandas==3.0.1
pathspec==1.0.4
pendulum==3.2.0
pluggy==1.6.0
prison==0.2.1
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
psutil==7.2.2
psycopg2-binary==2.9.11
py4j==0.10.9.7
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.12.1
pyogrio==0.12.1
pyproj==3.7.2
pyspark==3.5.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.2
rich-toolkit==0.19.7
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
setproctitle==1.3.7
setuptools==82.0.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.1
sqlparse==0.5.5
starlette==0.48.0
structlog==25.5.0
svcs==25.1.0
tabulate==0.10.0
tenacity==9.1.4
termcolor==3.3.0
text-unidecode==1.3
threadpoolctl==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==2.0.0
unicodecsv==0.14.1
universal_pathlib==0.2.6
urllib3==2.6.3
uuid6==2025.0.1
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.7
wheel==0.46.3
wirerope==1.0.0
wrapt==2.1.2
WTForms==3.2.1
yarl==1.23.0
zipp==3.23.0

